!Hotfix: RequestParameter 매핑 에러 이슈 및 @CrossOrigin 어노테이션 제거

See merge request s08-bigdata-recom-sub2/S08P22A503!206

Former-commit-id: 87aff4faa4602741580567471d31d6e3bd09daf9

+++ backend/src/main/java/com/a503/churros/service/news/NewsServiceImpl.java
@@ -97,7 +97,7 @@ public class NewsServiceImpl implements NewsService{
     }
 
     @Override
-    public ArticleDTO getArticleInfo(/*long userId ,*/ long articleId) {
+    public ArticleDTO getArticleInfo(/*long userId ,*/ Integer articleId) {
         Article article = ar.findFirstByIdx(articleId).orElse(null);
         if(article == null){
             return null;
